pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-agent-starter"
version = "0.1.0"
description = "General-purpose LangGraph agent scaffold for platform and algorithm teams"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.110",
  "gunicorn>=23.0",
  "uvicorn[standard]>=0.27",
  "dynamic-config-nacos>=0.1.0",
  "langgraph>=0.2",
  "langgraph-checkpoint-redis>=0.1",
  "langfuse>=3.14",
  "langchain>=0.3",
  "redis>=5.0",
  "httpx>=0.27",
  "tenacity>=8.2",
  "PyYAML>=6.0",
  "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
